[#42] chore : 채팅 학교 인증 redux 위치 변경
