+++ src/App.js
@@ -17,6 +17,7 @@ import MyPage from 'pages/MyPage';
 import ChatPage from 'pages/ChatPage';
 import ChatAuth from 'components/Chat/ChatAuth';
 import Unauth from 'components/Chat/Unauth';
+import ChatRoom from 'components/Chat/ChatRoom';
 
 const AuthorizedRoute = () => {
   const isLoggedIn = useSelector((state) => state.auth.isLoggedIn);
@@ -30,7 +31,7 @@ const AuthorizedRoute = () => {
 const App = () => {
   const dispatch = useDispatch();
   const isLoggedIn = useSelector((state) => state.auth.isLoggedIn);
-
+  const isSchoolAuth = useSelector((state) => state.myPage.isAuth);
   useEffect(() => {
     const token = getCookie('refresh_token');
     setTokenOnHeader(token);
@@ -59,7 +60,7 @@ const App = () => {
         <Route element={<AuthorizedRoute />}>
           <Route exact path="chat/*" element={<ChatPage />}>
             <Route path="auth" element={<ChatAuth />} />
-            <Route path="unauth" element={<Unauth />} />
+            <Route path="room" element={isSchoolAuth ? <ChatRoom /> : <Unauth />} />
           </Route>
         </Route>
       </Routes>
