+++ src/components/Chat/ChatRoom.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ChatRoom = () => {
+  return <div>인증됨</div>;
+};
+
+export default ChatRoom;
